Add files via upload
Speak and Translate is an android application which has been developed by me(Muhammad Umair Anjum). In this application i have tried my best to provide the ease to user who did not know about the multiple languages or they are staying abroad and don't know the native language their and also tried the ocr (by  extended my final year project in this domain) to provide the further ease to the user.